# docs/openapi.yaml
...
components:
...
    schemas:
...
        ChangeStatusRequest:
            properties:
                action:
                    enum:
                        - submit
                        - publish
                        - sell
                        - archive
                        - restore
                    type: string
            required:
                - action
            type: object
...
        CreateItemRequest:
            properties:
                attributes:
                    additionalProperties:
                        type: string
                    type: object
                description:
                    maxLength: 5000
                    type: string
                price:
                    minimum: 0
                    type: integer
                title:
                    maxLength: 200
                    minLength: 3
                    type: string
            required:
                - title
            type: object
...
        ErrorBody:
            properties:
                code:
                    type: string
                field:
                    type: string
                message:
                    type: string
                request_id:
                    type: string
            type: object
        ErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ErrorBody'
            type: object
...
        Item:
            properties:
                attributes:
                    additionalProperties:
                        type: string
                    type: object
                created_at:
                    type: string
                description:
                    type: string
                id:
                    type: string
                owner_id:
                    type: string
                price:
                    type: integer
                status:
                    type: string
                title:
                    type: string
                updated_at:
                    type: string
            type: object
...
        ItemListEntry:
            properties:
                created_at:
                    type: string
                id:
                    type: string
                owner_id:
                    type: string
                owner_name:
                    type: string
                price:
                    type: integer
                status:
                    type: string
                title:
                    type: string
            type: object
        ItemListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/ItemListEntry'
                    type: array
                    uniqueItems: false
                next_cursor:
                    type: string
            type: object
...
        UpdateItemRequest:
            properties:
                attributes:
                    additionalProperties:
                        type: string
                    type: object
                description:
                    maxLength: 5000
                    type: string
                price:
                    minimum: 0
                    type: integer
                title:
                    maxLength: 200
                    minLength: 3
                    type: string
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            description: |
                JWT, подписанный HS256. Клеймы: `sub` — UUID пользователя,
                `role` — роль, `exp` — момент истечения, `iat` — момент выпуска,
                `jti` — уникальный идентификатор токена.

                Просроченный токен и токен с неверной подписью одинаково
                дают 401 с кодом `unauthorized`.
            in: header
            name: Authorization
            scheme: bearer
            type: http
info:
    contact:
        name: Avito Hack Team
    description: |
        API объявлений с игровым слоем: пользователь заводит объявления, добавляет фотографии,
        ведёт избранное, а параллельно развивает питомца — за действия в маркетплейсе
        начисляется опыт, растут уровень и стрик, открываются награды.

        ## Конверт ошибок

        Все ошибки возвращаются единым конвертом `ErrorEnvelope`:

        ```json
        {"error": {"code": "validation_error", "message": "title must be at least 3 characters", "field": "title", "request_id": "..."}}
        ```

        Поля `field` и `request_id` опциональны и опускаются, когда пусты.
        Коды: `validation_error`, `bad_request`, `unauthorized`, `forbidden`,
        `not_found`, `conflict`, `internal_error`.

        ## Авторизация

        JWT (HS256) передаётся в заголовке `Authorization: Bearer <token>`.
        Срок жизни токена лежит в клейме `exp` внутри самого токена — отдельного поля
        `expires_at` в ответах на регистрацию и вход нет. В клеймах также присутствуют
        `sub` (идентификатор пользователя) и `role` (`user` | `moderator` | `admin`).
        WebSocket авторизуется query-параметром `token`, так как браузерный
        WebSocket API не позволяет задавать заголовки.

        ## Пагинация

        Списки используют keyset-курсоры, непрозрачные для клиента (base64url).
        В проекте два независимых формата курсора:

...

...

        ## Деньги

...

        ## Тело запроса

...
    license:
...
openapi: 3.1.0
paths:
...
    /api/v1/items:
        get:
            description: |-
                Анонимный доступ разрешён (`OptionalAuth`): токен читается, если передан,
                но не требуется. Анонимно и для чужого пользователя видны только
                объявления в статусах `published` и `sold`; владелец дополнительно видит
                собственные черновики.

                Пагинация — временной курсор. Побочных эффектов нет.
            operationId: listItems
            parameters:
                - description: Размер страницы. По умолчанию 20, максимум 100.
                  in: query
                  name: limit
                  schema:
                    default: 20
                    maximum: 100
                    type: integer
                - description: Временной курсор из `next_cursor` предыдущей страницы.
                  in: query
                  name: cursor
                  schema:
                    type: string
                - description: Фильтр по статусу.
                  in: query
                  name: status
                  schema:
                    enum:
                        - draft
                        - moderation
                        - published
                        - sold
                        - archived
                    type: string
                - description: Фильтр по владельцу. Невалидный UUID даёт 400.
                  in: query
                  name: owner_id
                  schema:
                    format: uuid
                    type: string
                - description: Полнотекстовый поиск по заголовку.
                  in: query
                  name: search
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ItemListResponse'
                    description: Страница объявлений
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
                - {}
            summary: Публичный список объявлений
            tags:
                - Items
        post:
            description: |-
                Создаёт объявление в статусе `draft` от имени владельца токена.

                Не идемпотентна — каждый вызов создаёт новую запись.

                Побочный эффект: публикуется доменное событие, на которое подписан
                игровой модуль. Опыт за создание объявления начисляется асинхронно
                (через Kafka либо через in-process шину, если Kafka выключена),
                поэтому в ответе изменение XP не отражается — его нужно получать
                из WebSocket-события `xp.gained` или повторным `GET /api/v1/pet`.
            operationId: createItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateItemRequest'
                description: Данные объявления
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Item'
                    description: Объявление создано
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Создание объявления
            tags:
                - Items
    /api/v1/items/{id}:
        get:
            description: |-
                Анонимный доступ разрешён. Чужой черновик или архив отдаётся как 404,
                а не 403, чтобы не раскрывать существование записи.
            operationId: getItem
            parameters:
                - description: Идентификатор объявления
                  in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Item'
                    description: Объявление
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
                - {}
            summary: Объявление по идентификатору
            tags:
                - Items
        patch:
            description: |-
                Частичное обновление: все поля тела необязательны, применяются только
                переданные. Доступно только владельцу — чужое объявление даёт 403.

                Идемпотентна: повтор с тем же телом приводит к тому же состоянию.
                Терминальные статусы (`sold`) редактировать нельзя — 409.
            operationId: updateItem
            parameters:
                - description: Идентификатор объявления
                  in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateItemRequest'
                description: Изменяемые поля
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Item'
                    description: Объявление обновлено
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Forbidden
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Conflict
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Редактирование объявления
            tags:
                - Items
...
    /api/v1/items/{id}/status:
        post:
            description: |-
                Переводит объявление по конечному автомату статусов. Действие задаётся
                полем `action`, а не целевым статусом.

                Разрешённые переходы:

                | Из | В |
                |---|---|
                | `draft` | `published`, `moderation`, `archived` |
                | `moderation` | `published`, `draft`, `archived` |
                | `published` | `sold`, `archived` |
                | `sold` | — (терминальный) |
                | `archived` | `draft` |

                Соответствие действий: `submit` → `moderation`, `publish` → `published`,
                `sell` → `sold`, `archive` → `archived`, `restore` → `draft`.

                Недопустимый переход даёт 409. Доступно только владельцу.

                Побочный эффект: публикация (`publish`) и продажа (`sell`) порождают
                доменные события, за которые игровой модуль асинхронно начисляет опыт
                питомцу.
            operationId: changeItemStatus
            parameters:
                - description: Идентификатор объявления
                  in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChangeStatusRequest'
                description: Действие
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Item'
                    description: Статус изменён
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Forbidden
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Переход не разрешён из текущего статуса
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Смена статуса объявления
            tags:
                - Items
    /api/v1/items/mine:
        get:
            description: |-
                Список объявлений владельца токена во всех статусах, включая черновики
                и архив. Фильтр `owner_id` недоступен — владелец подставляется из токена.
            operationId: listMyItems
            parameters:
                - description: Размер страницы. По умолчанию 20, максимум 100.
                  in: query
                  name: limit
                  schema:
                    default: 20
                    maximum: 100
                    type: integer
                - description: Временной курсор из `next_cursor` предыдущей страницы.
                  in: query
                  name: cursor
                  schema:
                    type: string
                - description: Фильтр по статусу.
                  in: query
                  name: status
                  schema:
                    enum:
                        - draft
                        - moderation
                        - published
                        - sold
                        - archived
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ItemListResponse'
                    description: Страница объявлений
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Мои объявления
            tags:
                - Items
...
servers:
...